requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
pydantic==2.9.0
python-multipart==0.0.9
pytest==8.3.2
httpx==0.27.0
tqdm==4.66.1